fix: Fix repetition in code and change the folder to which checkpoints are saved
There were two blocks of code doing the exact same thing - creating checkpoints.That could create a lot of confusion. For example, it can be a problem if someone wants to change the directory to which checkpoints are saved. Therefore, I deleted one of the blocks of code for creating checkpoints.
	I changed the directory to which checkpoints were saved for better organization of the repository. From now on all the checkpoints will be saved to a folder named training. The training folder is further divided into separate folders containing checkpoints for different algorithm implementations.